refactor(ui): split KeymapEditor.tsx into focused modules (1383→528 lines)
Extract the layout picker (state hook + presentational content), the
JSON-editor modal state (one generic helper behind an explicit
surface), the View Matrix wiring, the modal collection, the layer
keycode builders, the side toolbar, and the popover bridge into their
own modules, leaving the editor with composition and top-level layout.
Code moved verbatim; no test needed edits. Two review fixes along the
way: useUnlockGate now returns a stable object (its fresh literal was
silently defeating consumers' memoization) and the picker data shapes
moved to keymap-editor-types to break a type-only sibling cycle.
